refactor: move root-level modules into proper architecture layers
Move progress_store.py and error-formatting functions from utils.py into
the application/services/ layer where they belong in the hexagonal
architecture. Root progress_store.py now re-exports from the new location
for backward compatibility. Generic utilities (allowed_file,
parse_page_range_from_form, parse_plain_english_from_form) remain in
root utils.py. Remove dead save_timing_data() from routes.py since
DocumentProcessingService handles timing data saving.

- progress_store.py -> application/services/progress_store.py
  (with set_progress_store/get_progress_store for injectable instance)
- Error formatting functions -> application/services/error_formatting.py
- Update all imports in routes.py, document_service.py, and tests
