+++ application/services/document_service.py
@@ -8,16 +8,17 @@ import contextlib
 import json
 import logging
 from pathlib import Path
-from typing import Optional, TYPE_CHECKING
+from typing import TYPE_CHECKING, Optional
 
 if TYPE_CHECKING:
     from application.context.application_context import ApplicationContext
     from domain.models import PageRange, ProcessingResult, TimingMetadata
 
-from domain.models import ProcessingRequest, TimedAudioResult, TextSegment
+from application.services.error_formatting import clean_text_for_display, get_processing_stage_error
+from application.services.progress_store import cancel_operation, is_operation_cancelled, update_progress
 from domain.errors import ApplicationError, ErrorCode, Result
-from progress_store import update_progress, is_operation_cancelled, cancel_operation
-from utils import get_processing_stage_error, clean_text_for_display, parse_page_range_from_form, parse_plain_english_from_form
+from domain.models import ProcessingRequest, TextSegment, TimedAudioResult
+from utils import parse_page_range_from_form, parse_plain_english_from_form
 
 logger = logging.getLogger(__name__)
