+++ routes.py
@@ -21,18 +21,20 @@ from werkzeug.utils import secure_filename
 from application.config.system_config import SystemConfig
 from application.context.application_context import ApplicationContext
 from application.services.document_service import DocumentProcessingService
-from domain.container.service_container import ServiceContainer
-from domain.models import PageRange, ProcessingResult
-from infrastructure.file.file_manager import FileManager
-from progress_store import (
+from application.services.error_formatting import (
+    get_contextual_error_message,
+    get_processing_stage_error,
+)
+from application.services.progress_store import (
     cancel_operation,
     get_progress,
     update_progress,
 )
+from domain.container.service_container import ServiceContainer
+from domain.models import PageRange, ProcessingResult
+from infrastructure.file.file_manager import FileManager
 from utils import (
     allowed_file,
-    get_contextual_error_message,
-    get_processing_stage_error,
 )
